+++ common/tests/codegen-tests/playwright-typescript/playwright-typescript-gen.test.ts
@@ -1,18 +1,18 @@
 import os from "os";
 import fs from "fs";
 import path from "path";
-import { createTestDataGeneral } from "./playwright-typescript-general.test-data";
 import { createRmTestProject, writeOutputProjectFiles } from "../../test-helpers/rm-project-generator";
 import { PlaywrightTypeScriptProjMetaGenerator } from "../../../src/codegen/playwright-typescript/playwrightTypeScriptMetaGenerator";
 import { createTempDir } from "../../test-helpers/fsHelpers";
 import { generateCode } from "../../../src/codegen";
 import { IProgressEvent } from "../../../src";
 import { execSync } from "child_process";
+import { createTestDataKitchenSink } from "../playwright-csharp-common/playwright-common.test-data";
 const tempDir = path.resolve(__dirname, "../../.tmp");
 
 test("Generate Test project Playwright Typescript General", async () => {
   // Arrange
-  const projSpec = createTestDataGeneral();
+  const projSpec = createTestDataKitchenSink();
   const tmpDir = createTempDir("playwright-typescript-general");
   const copyToDir = path.join(tmpDir, "rmproj");
   fs.mkdirSync(copyToDir);
