+++ common/tests/codegen-tests/playwright-typescript/playwright-typescript-meta.test.ts
@@ -3,9 +3,11 @@ import fs from "fs";
 import path from "path";
 import { createTestData } from "./playwright-typescript-meta.test-data";
 import { createRmTestProject } from "../../test-helpers/rm-project-generator";
-import { PlaywrightTypeScriptProjMeta } from "../../../src/codegen/playwright-typescript/playwrightTypeScriptMeta";
 import { createSourceProjectMetadata } from "../../../src/codegen/codegenSourceProjectMeta";
 import { createTempDir } from "../../test-helpers/fsHelpers";
+import { PlaywrightTypeScriptProjMetaGenerator } from "../../../src/codegen/playwright-typescript/playwrightTypeScriptMetaGenerator";
+import { IProgressEvent } from "../../../src";
+import { generateCode } from "../../../src/codegen";
 
 test("Generate output projec meta Playwright Typescript", async () => {
   const projSpec = createTestData();
@@ -14,8 +16,14 @@ test("Generate output projec meta Playwright Typescript", async () => {
   const copyToDir = path.join(tmpDir, "rmproj");
   fs.mkdirSync(copyToDir);
   const projFile = createRmTestProject(projSpec, copyToDir);
+
   const sourceProjMeta = await createSourceProjectMetadata(projFile);
-  const outputMetadataProcessor = new PlaywrightTypeScriptProjMeta(sourceProjMeta);
+
+  await generateCode(projFile, (event: IProgressEvent) => {
+    // console.log(`PROGRESS: ${event.type}. ${event.log}`);
+  });
+
+  const outputMetadataProcessor = new PlaywrightTypeScriptProjMetaGenerator(sourceProjMeta);
   const outputProjMetadata = outputMetadataProcessor.createOutputProjectMetadata();
   expect(outputMetadataProcessor).toBeTruthy();
 });
